feat(images): add image upload for users

+++ src/Service/FileUploaderService.php
@@ -10,17 +10,31 @@ use Symfony\Component\String\Slugger\SluggerInterface;
 
 class FileUploaderService
 {
+    private string $baseUploadDirectory;
+    private SluggerInterface $slugger;
+
+    private const UPLOAD_DIRECTORIES = [
+        'questions' => '/uploads/images/questions',
+        'users' => '/uploads/images/users',
+    ];
+
     public function __construct(
-        private string           $targetDirectory,
-        private SluggerInterface $slugger,
+        string           $baseUploadDirectory,
+        SluggerInterface $slugger,
     )
     {
+        $this->baseUploadDirectory = $baseUploadDirectory;
+        $this->slugger = $slugger;
     }
 
-    public function upload(UploadedFile $file, ?string $oldFilename = null): string
+    public function upload(UploadedFile $file, string $entityType = 'questions', ?string $oldFilename = null): string
     {
+        if (!array_key_exists($entityType, self::UPLOAD_DIRECTORIES)) {
+            throw new Exception("Invalid entity type: $entityType");
+        }
+
         if ($oldFilename) {
-            $this->delete($oldFilename);
+            $this->delete($oldFilename, $entityType);
         }
 
         $originalFilename = pathinfo($file->getClientOriginalName(), PATHINFO_FILENAME);
@@ -28,21 +42,26 @@ class FileUploaderService
         $fileName = $safeFilename . '-' . uniqid() . '.' . $file->guessExtension();
 
         try {
-            $file->move($this->getTargetDirectory(), $fileName);
+            $file->move($this->getTargetDirectory($entityType), $fileName);
         } catch (FileException $e) {
             throw new Exception('Error uploading file');
         }
 
+
         return $fileName;
     }
 
-    public function delete(string $filename): bool
+    public function delete(string $filename, string $entityType = 'questions'): bool
     {
         if (!$filename) {
             return false;
         }
 
-        $filePath = $this->getTargetDirectory() . '/' . $filename;
+        if (!array_key_exists($entityType, self::UPLOAD_DIRECTORIES)) {
+            throw new Exception("Invalid entity type: $entityType");
+        }
+
+        $filePath = $this->getTargetDirectory($entityType) . '/' . $filename;
 
         if (file_exists($filePath)) {
             return unlink($filePath);
@@ -51,8 +70,12 @@ class FileUploaderService
         return false;
     }
 
-    public function getTargetDirectory(): string
+    public function getTargetDirectory(string $entityType = 'questions'): string
     {
-        return $this->targetDirectory;
+        if (!array_key_exists($entityType, self::UPLOAD_DIRECTORIES)) {
+            throw new Exception("Invalid entity type: $entityType");
+        }
+
+        return $this->baseUploadDirectory . self::UPLOAD_DIRECTORIES[$entityType];
     }
 }
